Further build cleanup.
The build system used 'LINUX' as a shorthand for Unix-like systems, which
ends up being a horrible mess -- it's pretty counterintuitive to have
the Solaris/Irix/BSD setups #defining LINUX to make things work.

This replaces the use of 'LINUX' with a more-general and more-obvious
'UNIX', and moves the not-really-linux-at-all 'liblinux' to 'libunix'.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@6 c06c8d41-db1a-0410-9941-cceddc491573