License compliance changes -- see the following excerpt from license.txt:
	2. You may modify your copy or copies of Crawl or any portion of it,
	   and copy and distribute such modifications under the terms of
	   Paragraph 1 above (including distributing this Licence Agreement),
	   provided that you also do the following:
	     a. cause the modified files to carry prominent notices stating
	        that you changed the files and the date of any change; and

This tags all source files that didn't already have it with a 'Modified for
Crawl Reference by $Author on $Date$' header tag, allowing SVN to do the
datestamp monkeying for us.

I've also tagged files that are new to crawl-ref / stone soup, which the
license doesn't require but does provide a bit of unity and makes it easier
to see who's changing things.



git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/branches/stone_soup@124 c06c8d41-db1a-0410-9941-cceddc491573