[1601595] Lava smokes, swamp shallow water tends to mist.
Cold spells fired over water trail freezing cloud.

Smoke/steam spreads out a bit as it dissipates. I can't seem to make this look
as pretty as I imagined it. :-(

MAX_CLOUDS is still on 100. We may want to increase this.

Monster pale draconians are pegged back to speed 10 (was 12).

git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@486 c06c8d41-db1a-0410-9941-cceddc491573