Fixed Abyss banishment crash.
Don't show -more- after the "cast into the Abyss" so that players on public 
servers are not tempted to kill the tty to escape the banishment. It's still
possible that there are enough messages in view that the game puts in a -more-
unasked for, but I'll live with that.

git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@769 c06c8d41-db1a-0410-9941-cceddc491573