[1631021] Interrupt delays only if the player knows he was attacked.
Under the old system, the delay was interrupted when the monster got an attack
round against the player; if the monster is invisible and misses, the player
received no message, but delays were interrupted.

Under the new system, the delay is interrupted only if the player perceives the
attempted attack (if the player shield-blocks, or the attacker is visible, or
the player actually takes a hit).

Also fixed the old behaviour where you would not be aware of an unseen horror
if it hits but does no damage.

git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@816 c06c8d41-db1a-0410-9941-cceddc491573