Stage 1 of the direction() rewrite. I am committing this even though it isn't complete yet, because: (a) it's good enough now that I like it better than the old version (b) feedback and testing is badly needed.
Direction selection has been redone and is no longer modal. You now
use the normal arrow or vi keys to move around. If you want to select
a direction, as opposed to a specific target square, use the Shift-
versions of the movement keys. This might not work with arrow keys
on some systems; it's hard for me to check on my own system.

Most of the other keys should work, but I'm not sure I've got them all.
';', '+', '-', '.', ' ' (== '.'), 't', 'f', 'p' all work.

'x' (from the main screen) still calls look_around(), not the new
direction code, so you have some room to compare.

The confirm_* options are now meaningless and will be removed shortly.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@880 c06c8d41-db1a-0410-9941-cceddc491573