Interlevel travel now handles branch stairs more intelligently:
Travel knows that all branch exits of the same type share the same destination,
and will use the nearest branch exit, even if you've never used it before.
 
The same applies to multiple branch entries, the portals to Zot:1 for instance.
This does mean that the first visit to Zot is likely to generate a bad Zot-exit
guess (Zot has many entries and many exits), but that'll be fixed the first
time the player leaves Zot.

.tc version updated.

git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@945 c06c8d41-db1a-0410-9941-cceddc491573