More branch entry vaults (Erik).
Changed behaviour when using KMONS and KITEM without a corresponding KFEAT.
Previously this would force the dungeon feature of that square to plain floor.
The new behaviour is to try to map the glyph to itself. For instance:
KMONS: O = pandemonium demon
will place a pandemonium demon on the rune, whereas previously it would place
the pan demon on a floor square.

Tossed out old waypoint hacks. Waypoints are now handled in the same way as
travel to stash search places. This ought to fix the waypoint corruption issues.

git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@1081 c06c8d41-db1a-0410-9941-cceddc491573