Vampires, yay! Credit mostly goes to Jarmo, though there are a few traces of things I've changed (that I've included and then commented out).
There's still lots of stuff to be added, but they 
should be playable right now. I notice that their 
vampiric bite attack needs to be made more
probable. Feel free to comment on anything that 
looks weird, is plain wrong, or goes against the 
spirit of Crawl/Stonesoup. 
Positive feedback will be appreciated as well. :)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@1870 c06c8d41-db1a-0410-9941-cceddc491573