In my previous commit I also moved equipment slots around, so the spears of Berserkers start at 'c'. Now I did the same for Merfolk Hunters (javelins).
Also: removed enforced random weapon for Chaos Knights
(FR 1772666). I am very much aware of the reason behind this
rule (chaos = randomization), but a) it probably _does_ lead 
to startscumming, b) we do not wildly randomize gods or 
stats, and c) the weapons are not important enough to really 
enforce this. All IMHO, of course.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2050 c06c8d41-db1a-0410-9941-cceddc491573