Added two command hints: 'v'iewing for certain features, and how to turn autopickup/autoprayer on again after an encounter with an invisible creature.
I also modified selection of book shops, so the offered 
books are more likely to be different. This changes 
balance in that on the whole more books are present, so
maybe we should simply throw duplicates out rather than
reroll. Oh, and manuals are now even rarer than before.
Maybe they should be excepted from this new rule.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2056 c06c8d41-db1a-0410-9941-cceddc491573