Fix 1708493: corpses being butchered turning into skeletons, yet not, and leaving clean chunks. Strange stuff!
I also added a restriction to build_river() and build_lake()
so that staircases or branch entries/exits don't drown. This
might still result in stairs in the midst of a lake but that
is vastly better than no stairs at all, as probably happened
with 1787377. Now I don't know whether this was even played 
a part in that occurence, but I didn't find anything else in
dungeon.cc that might have caused it.

And YAY, I'm a member!


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2060 c06c8d41-db1a-0410-9941-cceddc491573