Changed requirements for auto-identification of equipped items, specifically jewellery and randarts. The goal of this change was to clear up the identification rules and apply them consistently. It was inspired by FR 1788698.
* Like the ring of teleportation, =lev and "rage now only
  autoID if you do not wear/wield a randart with the same
  property.
* Stat changes caused by randarts always give a message.
* Other "obvious" changes give a message only if the 
  randart has not been identified, to avoid message spam.
  (Obvious refers to attributes that would cause the item
   to be autoID'd if it were the appropriate sub type.)
* Randarts do NOT autoID anymore just because you recognize
  the base type.

On the whole, autoidentification becomes much rarer (thus 
both the scroll and the spell become more valuable). Yes,
this might be too harsh. On the upside, the player is told 
about properties such as stat changes (including AC and EV) 
and all evokable abilities. 
We might even include information about items hindering or 
preventing spellcasting and teleportation, respectively, 
under the assumption that skilled spellcasters and chars 
with teleportitis or teleport control would notice.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2067 c06c8d41-db1a-0410-9941-cceddc491573