Changed the look of bazaars.
Here, walls will always be golden (yellow) and the floor
will be any of blue, red, lightgreen, lightblue, magenta or
green, depending on the dungeon depth (modulo 6). This
might be a bit weird, but guarantees the same colour
regardless how often the map is redrawn. Also, bazaars 
should be sufficiently rare for that not to matter.

Removed all SHUFFLE statements in bazaar.des about
wall type - it's all rock now.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2072 c06c8d41-db1a-0410-9941-cceddc491573