HANDS_DOUBLE -> HANDS_TWO (rather than HANDS_HALF). I hope I found the correct place to change it. Also does this need to be balanced somehow?
Paralysed monsters are now stabbable. Anything else 
doesn't make a lot of sense. Now this might make
paralysis too strong, so maybe we need to up the MR 
threshold and/or make the spell more expensive.

Also fixed a minor bug that had paralysed monsters
changing places when leaving the level for a short enough
period for the enchantment to still be in place.

Conjure flame cannot be placed on the player's position.
(Bug 1793587).


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2090 c06c8d41-db1a-0410-9941-cceddc491573