Applied Enne's fix to 1795522: "show_beam shouldn't work when just looking..."
While I was at it, I also added another parameter 
needs_path that is set to false for spells that can be 
applied directly to a target (SPFLAG_GRID, SPFLAG_TARGET)
such as Smiting, Conjure Flame or Freezing Cloud, so for 
these the initial beam path doesn't show, even if the
option should be set to true.
Note "initial beam path" - you can still toggle it to 
show the path anyway. For cases such as this we might
enforce not drawing the path but there would have to be
some kind of feedback, so the player doesn't go crazy
trying to toggle the option. ;)

I'll probably do it later, once I've found a way to give 
the information. Or would that be too spoily, in any case?


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2091 c06c8d41-db1a-0410-9941-cceddc491573