Yet another change to monster speech and silencing...
I found my previous change to block all messages passed 
through MSGCH_TALK when silenced was too restrictive, so
I've added another channel MSGCH_TALK_VISUAL and adapted the
monster speech files accordingly. For example, imps can now
also "talk" when silenced, but only the visual stuff will
actually be output. (This automatically takes care of 
lowering the output likelihood.) And no, this doesn't hurt 
any, as visual talk also won't interrupt travel or resting.

And I added a how-to file on monster speech, which was the
main reason for my silence during the last couple of hours. 
It's not that the database system is that complicated but it 
surely can't hurt to give players who want to add something 
an idea of what is possible and how to go about it.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2108 c06c8d41-db1a-0410-9941-cceddc491573