A handful of new game options:
menu_colour_prefix_id, if set to true, causes the identification state
of an object to be prefixed to the string menu colour regexes are
matched against (but does not alter the string displayed on the screen).

menu_colour_prefix_class, if set to true, prefixes the object's base type
to the string menu colour regexes are matched against.

user_note_prefix can be set to a string which will be prefixed to
manual user notes when they are displayed, to make them easier to find.

detailed_hunger, if set to true, will cause three new informational-only
hunger states to be displayed: near starving, very hungry and very full.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2121 c06c8d41-db1a-0410-9941-cceddc491573