Added additional optional parameter to item_def::name(), ignore_flags, which will cause the name to be constructed as if those item flags had been unset.
Give an auxiliary cause of death for a stat going below 1.  (Death
by stat loss is already pretty rare, and death by stat loss with
confusion as to what caused the stat loss must be *really* rare, but still,
if you were confused about what caused the stat loss that lead to death,
that'd be pretty frustrating)

ouch() is now called from within modify_stat() and lose_stat() right after
the stat is lowered, rather than when the stat is updated on the screen.
This incidentally fixes the minor annoyance of saying "no" to stat loss
death in wizard mode, only to be asked if you want to die every time
the screen is updated until you fix having a non-positive stat.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2179 c06c8d41-db1a-0410-9941-cceddc491573