?/ now asks if you want to describe a monster, spell, or feature, and filters out matches if they aren't of the desired type.
If there's more than one match, after selecting a match to look at,
exiting from the description will return you to the menu, rather than
to the dungeon.

If you've asked for monsters, you can toggle sorting of the menu
between alphabetical and by aproximated monster toughness (this probably
still needs some work, since it seems to say that ordinary worms are
tougher than brain worms).

Only the monster symbol is coloured when showing a menu of monsters
to describe.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2185 c06c8d41-db1a-0410-9941-cceddc491573