This commit breaks save file compatability.
Lots of new things that amuse/stimulate Xom, and a few things which
don't amuse him as much anymore.  Among the new things is a corpse
turning into a skeleton while butchering it; if this is too harsh to
do just for Xom's amusement (previously turning into a skeleton while
butchering was an ignored case and still produced chunks of flesh) it
can be changed back.  Also, if a Xom worshiper draws the Blank card,
Xom makes it act like a Xom card, since a plain old Blank card is boring.

Keep track of which branch the Orb is in, if the player isn't carrying
it.

Keep track of how/why the player ended up in a particular level type
(Abyss, Pan, etc).

Changed most "a distortion effect" cause strings for distortion caused
tranlsocation miscast effects to something more specific.

Added new wizard commands 'C' to curse or uncruse an item, and 'Ctrl-A'
to re-generate the Abyss.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2256 c06c8d41-db1a-0410-9941-cceddc491573