Implementing merfolk monsters. There are two types, merfolk fighters, and mermaids. The latter are more interesting. ;)
Whenever a mermaid sings there's a chance you get beheld, meaning you
cannot move away from the mermaid, that is cannot increase the 
distance between you and her. (Because of the distance pecularities
this means at dist 1 you can only be NSW or E of the mermaid but not
the diagonal directions.) 
If you manage to be beheld by several mermaids at the same time, your 
movements have to respect the distance to all of them.

I've added a vector beheld_by that keeps track of beholding monsters
and makes checking distance and updating beholding status easier.

Merfolk are immune to the mermaid song.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2453 c06c8d41-db1a-0410-9941-cceddc491573