If a submerged monster moves onto a grid in which it can't be submerged, force it to unsuberge (i.e., a mermaid moving directly from water onto floor without first unsubmerign).
If a submerged monster shouts, force it to unsubmerge first.

If a visible submerged monster unsubmerges within view of the player,
say that it "bursts forth from the water" if it can travel over land,
and that it "surfaces" otherwise, rather than saying that it "comes
into view".  If an invisble sumberged monster unsubmerges within view
of a player who can't see invisible, and it can move over land,
say "something invisble bursts forth from the water".


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2466 c06c8d41-db1a-0410-9941-cceddc491573