Another mutation commit. Sorry, I couldn't resist.
I'm replacing the (stupid, I agree) "extra eyes" mutation 
with the more Crawlific "shaggy fur" mutation. It currently 
doesn't count towards the scales counter, but that could be 
changed. Also, maybe it could replace one of the scales 
with a similar AC bonus. Oh, and trolls start out with this 
at level 1 because I think it fits them.

Adding the percentage bonus (rarity 1) and penalty (rarity 9)
for mana. Maybe the bonus (+10%, +20%, +30%) should be 
abolished entirely as David suggested; for now I've left it 
as a mutation pair.

As a third change I'm including the experimental sleepiness
mutation that makes you randomly fall asleep. Being hit, 
loud noises and starving will make you wake up.

Again, may break mutations of existing save files.

---------------------------------

To answer David's question on balancing (or rather, make 
everything more complicated, I'm afraid) I've counted 
mutations in 0.3 and as of this commit, to compare them. 

Rarity influences the chance a mutation, once randomly
chosen, will be considered acceptable, with a chance of 
rarity/10. Each round of trying to find an acceptable 
mutation there's a 1/1000 chance of just giving up. 
In brackets I will list the average probability of a 
good/bad mutation being chosen in a given round. 
(The difference to 100 is the chance of having to reroll.)

in 0.3:
46 good mutations of average rarity 3.2 (23.72%)
12 bad mutations of average rarity 8 (15.47%)
4 in-between mutations of average rarity 7 (4.51%)
(56.3% chance of rerolling)

new:
48 good mutations of average rarity 3.1 (21.72%)
14 bad mutations of average rarity 7.6 (15.71%)
6 in-between mutations of average rarity 5 (4.41%)
(58.16% chance of rerolling)

By in-between mutations I mean ones with advantages and
drawbacks, namely weak_flexible, strong_stiff, horns, 
hooves, talons and paws. (The last four are probably seen
as mostly negative by a lot of players due to losing an
equipment slot.)

Because of the lower rarity of the new mutations (including 
drifting) the overall chance of _not_ receiving a mutation
has slightly increased (1/1000 chance per reroll necessary).


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2640 c06c8d41-db1a-0410-9941-cceddc491573