Add consistency fixes for elven weapons:
* Since monster elves can get elven scimitars, and scimitars are long 
blades and better for the dexterous (both of which fit elves), allow 
elven scimitars to be randomly generated.  The chance is 1/5, falling 
through to the existing case of 1/2 for orcish ones, based on the chance 
for elven long swords and adjusted for scimitars' greater rarity.

* Since, aside from starting equipment for elf players, no elven axes 
show up, either randomly or on monster elves, the elven aptitudes for 
axes are almost as bad as those for the disallowed maces/flails, and 
axes are better for the strong (none of which fit elves), disallow elven 
axes.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@2956 c06c8d41-db1a-0410-9941-cceddc491573