Fixed some randart auto-inscription bugs.
For jewellery randarts, the properties of the randart's basetype are
included in the auto-inscription.

Auto-inscribe unidentified randarts if the the player has learned
some of its properties via equiping it.

If an unidentified randart is autoinscribed, is subsequently identified,
and the identification reveals more of its properties, then remove the
old atuo-inscription string before adding the new one.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3000 c06c8d41-db1a-0410-9941-cceddc491573