Replace TSO's Smiting with Divine Shield.
Actually, I wasn't entirely sure which of the many attack
invocations to replace and I chose Smiting mostly because
of the new overlap with Zin and because such a change is
easiest to revert (copy code from Zin). :p

Divine Shield strengthens existing shields (also magical,
which might have to be excluded) or else creates a new
magical one. Either way, the shields value is increased by
  3 + skills(SHIELDS)/5
and this bonus lasts for 
  5 + (skills(SHIELS) + 2*skills(INVOCATIONS))/6
turns, after which it is decreased by 1 every two out of
three turns until it reaches 0.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3225 c06c8d41-db1a-0410-9941-cceddc491573