Add gore and splatter to Crawl - in the form of Jarmo's (strongly modified, I admit) blood patch.
Whenever the player or a monster capable of bleeding suffers 
damage (currently only in melee) or when you dissect a corpse
there's a chance (depending on damage/corpse weight) that the
square will turn bloody and be coloured red, with a much 
lower chance of additionally spattering some of the 
surrounding squares. These chances are probably still a bit 
too high for later values of possible damage dealt.
Also, ranged and some of the magic attacks (earth-based, 
mostly) should also be able to send blood flying. :p

For now, this is flavour only, but it shouldn't be too 
difficult to introduce relations to Necromancy or evil gods.

Also consider dragons to be warm-blooded, like dinosaurs.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3230 c06c8d41-db1a-0410-9941-cceddc491573