Add various crossbow-related fixes:
Fix [1864360]: Reenable random generation of crossbow bolts, as it was 
(erroneously?) removed in r1732, when javelins and sling bullets were 
introduced.  The probability for bolts is slightly greater than that of 
needles (12 vs. 10), as the probability of randomly generated crossbows 
is slightly greater than that of blowguns (6 vs. 5).  Furthermore, the 
probability for bolts is lower than the probability for arrows (12 vs. 
20), since the former do more damage.

Orc knights and warlords that get crossbows now get orcish ones instead 
of non-racial ones.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3315 c06c8d41-db1a-0410-9941-cceddc491573