Add mouseclick handling for menus in tiles: L-click will page down, R-click is treated like Escape.
Change L-click on player from single turn search to (the more
useful, I think) pickup, and add L-click to list of 
alternatives to call the pickup menu, so that clicking on the 
player icon twice will directly open the pickup menu.

This is getting complicated: there are so few combinations
of mouseclicks possible, but because they do different things
under different circumstances this *really* needs to be
documented somehow.

Fix coloured hp/mp bars not updating during running. (During 
travel the map needs to be redrawn anyway, so also having to 
redraw the bars doesn't hurt any.)

Also fix list_rotten to be readable from init.txt (and default 
to true).


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3316 c06c8d41-db1a-0410-9941-cceddc491573