Remove GLAMOUR, and introduce RECITE (new Zin invocation)
The latter is a work in progress and may have strange results.
It is intended as a multiturn invocation that is interrupted by
large hp losses. IMHO, reciting things is an action that needs
to take several turns. Each turn an attempt at influencing
intelligent monsters is made. They usually fall asleep, are
confused or turn berserk. 
The delay gets cancelled when the player is heavily wounded or
when there are no impressionable monsters around anymore.

I think the power formula will have to be tweaked.

Comments welcome!


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3321 c06c8d41-db1a-0410-9941-cceddc491573