Add shaft tile. Yay!
I've set this as tile for DNGN_TRAP_NATURAL in general as the shaft is 
currently the only one among those. This may need to be adapted later on. 
Ideally, the different kinds of mechanical traps (most notably blade and 
net) should get a tile of their own rather than use the default (arrow) 
one.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3350 c06c8d41-db1a-0410-9941-cceddc491573