Change monster glyphs as suggested by Markus Maier (Monster glyph adjustments v2, of 2008-02-12), with one small exception: I deemed yellow too bright for quokkas (yellow being the DANGER! sign of, among others, killer bees, death yaks, and Sigmund) and chose white instead.
For a complete list of changes see docs/monster_glyphs.txt,
which also doubles as options setting for init.txt.

I wasn't sure what default path to choose for the include
option, so it'll probably have to be adjusted for it to
work. (The message history shows an error message if a 
file to-be-included can't be read in.)


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3472 c06c8d41-db1a-0410-9941-cceddc491573