[FR 1903593] Change identification process of              ?recharging, ?enchant armour and ?identify.
If you read one of these scrolls and the type isn't known yet you are
prompted with "Modify which item?" (better message needed!) and get to
choose from the entire inventory.
If the chosen item can be usefully "modified" by the scroll (unID'd item
for identify, wand for recharging, enchantable armour for EA) the usual
effect takes place and the scroll is identified. (Reading other scrolls 
of the same type will then only offer a more sensible selection of items.) 
Otherwise, nothing happens.

Further, recharging and enchant armour now allow direct choice of the 
item in question, and it doesn't even have to be wielded or worn.

I think this change actually makes the id game more interesting and also
improves the interface. Gameplay might dictate that we reintroduce the
"armour needs to be worn" rule, but that remains to be seen until after 
some more playtesting.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3514 c06c8d41-db1a-0410-9941-cceddc491573