Tweak Vampires, part I
Better stats and aptitudes as suggested in David's post "The plan"

Minor modifications to starving/hungry/full messages. Generally change
Hungry to Thirsty, and Engorged to Alive. For now, the rest remains as
for the other species. Yes, this is nowhere near as radical as David'd 
suggestions as those are confusing me immenseley, and I just can't code 
that way.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3526 c06c8d41-db1a-0410-9941-cceddc491573