Fix 1909592: trunk not compiling on Windows
Improve information output for vampires in the screens 'A', '@' and '%'.
Reduce the "in touch with the powers of death" effect to one boost at xl 6,
and disallow bat form at engorged. Also, modify messages if they are not
available right away for feeding reasons.

What's still missing is a way to output those temporary resistances 
(currently only discernible from the '%' screen), preferentially in the
form of some general statements on 'A': "When hungry, you gain resistance
to cold." and more detailed descriptions on '@'.

Fix a couple of wrong hunger status checks.

Comment out the "Very agile monsters are quite hard to hit in your current 
equipment." messages because to my knowledge they hold little importance
right now. Please correct me if that view is wrong.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3534 c06c8d41-db1a-0410-9941-cceddc491573