Some more monster speech additions:  * allow handling of specifically neutral or hostile monsters  * use "good god" or "evil god" prefix where applicable  * add speech for friendly/neutral holy beings in combination with good god    followers
And... rewrote large portions of monster_speech.txt. I think I may now
finally have understood how this works, so in the not so distant I will have
have to clean up some of the hopelessly complicated functions I introduced 
when I only _thought_ I'd grasped the details. As usual, explaining it to 
someone else (if only in writing the documentation) really helped.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3591 c06c8d41-db1a-0410-9941-cceddc491573