More progress on TSO.  His retribution is now as follows:
* 33% chance of summoning daevas.

* 33% chance of blasting you with cleansing flame.  If you worship an 
evil god, you are now vulnerable to cleansing flame even if you're not 
undead or demonic; it seems fitting, especially as many living evil 
monsters seem to be priestly types.  (The code for the blast is 
currently based on a modified version of Xom's divine lightning.)

* 33% chance of booming voice.  This is currently identical to Zin's 
version, except that the message is different.

His smiting has been removed.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3593 c06c8d41-db1a-0410-9941-cceddc491573