Adjust TSO retribution probabilities.  New rules (not in comments yet because they may be adjusted again):
3/7: summon daevas
2/7: cleansing flame
2/7: 50% chance of noise, 50% chance of silence (anti-noise)

The original proposal also mentioned pestilence as an option, probably 
to be used on low-level characters instead of daevas, but the noise 
already overlaps with Zin (and fits, I think, given TSO's anti-stealth 
stance), and adding pestilence would just overlap with Zin more.  Maybe 
there should be a different *kind* of pestilence for TSO, or some other 
low-level summoning that would fit him?


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3602 c06c8d41-db1a-0410-9941-cceddc491573