Morphed SingleFileDB into TextDB, a class that accepts any number of input files.  Use it for all the databases, at least all the ones managed by database.cc.  Gets rid of 150 lines of code :)
Also, fix 1910634 by adding insult.txt to the shout db.  It results in
some duplication in the dbs on disk, but it's an OK fix for now, I think.

The DBs are now in the subdirectory saves/db/ ; is this a problem for
distribution?  I poke around in that directory occasionally and it's
convenient to have the DBs out of the way.



git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3680 c06c8d41-db1a-0410-9941-cceddc491573