Various Xom neutrality adjustments:
* Allow him to randomly summon angels or daevas (on the grounds that 
they're misguided, curious, exiled from their usual god, etc.), with the 
same rarity as greater demons.  Add appropriate introductory messages 
for this (from both me and dploog).

* Remove "Xom the Psychopath" from his list of titles, and replace it 
with the older "Xom the Begetter of Discontinuities" for now.

* Remove "Call Imp" from his list of spells that he can randomly make 
you cast, as he seems to summon enough minor demons through other means, 
which makes it an overlap (as the Summon Guardian and Summon Daeva 
"spells" would also be).


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3696 c06c8d41-db1a-0410-9941-cceddc491573