Another one of those mixed commits.
* Fix 1915277: Blade Hands time-out removing bardings.
* Ely's Lesser Healing costs range(0,1) piety (avg. 0.5)
* Allow synonym for "yes" in yesnoquit(), and use it for eating (e=y).

Also, in preparation for 0.4 (yes, I can't wait) finally disallow random
generation of weapons of orc slaying. Demon lords use a check of their
own and as a randart property it's also allowed, which I think is fine.
Allow random generation of nets and increase shop prices for nets and
javelins.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3766 c06c8d41-db1a-0410-9941-cceddc491573