Add fountains of blood (as suggested by Lemuel), and tidy up the existing fountains: change their order and  remove all those superfluous dry fountain types.
I couldn't find any place for random generation of fountains,
so if that is possible it will have to be disabled for blood
fountains outside the hells. I've added a shortcut (Y) for
vault creation, though. :)
Drinking from such a fountain has a 33% chance of leaving it
dry. I think that's it... for now.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3886 c06c8d41-db1a-0410-9941-cceddc491573