Finally allow ranged attacks (BEAM_MISSILE) and spells (BEAM_MMISSILE only) to also saturate the blood with floor.
Add a basic blood scent effect: all hounds, vampires, and the mosquito get
M_BLOOD_SCENT which will let them wake up when blood is spilled in their
vicinity. smell_blood() is rather similar to noisy(), so monsters will try
to seek the origin of the blood spilled.

Player vampires now get a message if blood is spilled outside their LOS,
"You smell fresh blood" with a distance estimate ("near-by", "in the distance")
added for vampires that are (near) bloodless. Also, the radius a vampire will
be capable of smelling blood will increase the thirstier he is.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3936 c06c8d41-db1a-0410-9941-cceddc491573